Move asset/public files to shared "assets" directory
Assets will not be moved to `dist/` during compilation.
server_IDs.json also moved to this so imports can reference a shared
path.
