+++ assets/slotpatterns.png
