+++ src/db/models/User.ts
@@ -1,7 +1,7 @@
 import mongoose, { Schema, model, type Model } from 'mongoose';
 import mongooseLong from 'mongoose-long';
+import servers from '../../../assets/server_IDs.json' assert { type: 'json' };
 import { STARTING_GOLD } from '../../constants/constants.js';
-import servers from '../../server_IDs.json' assert { type: 'json' };
 import {
     calculateDouggedPercentage,
     checkGoldClaimAvailability,
