+++ src/embeds/slot_patterns_embed.ts
@@ -1,9 +1,8 @@
 import { AttachmentBuilder, EmbedBuilder } from 'discord.js';
 import { join as pathJoin } from 'path';
-import { getDirName } from '../util/dirname.js';
 
 export const image = new AttachmentBuilder(
-    pathJoin(getDirName(import.meta.url), '..', 'images/slotpatterns.png')
+    pathJoin(process.cwd(), 'assets/slotpatterns.png')
 );
 export const patterns = new EmbedBuilder()
     .setColor(0xa51c30)
