+++ src/members/handle_members.ts
@@ -1,6 +1,6 @@
 import { Guild, GuildMember, PartialGuildMember } from 'discord.js';
+import servers from '../../assets/server_IDs.json' assert { type: 'json' };
 import { User } from '../db/models/User.js';
-import servers from '../server_IDs.json' assert { type: 'json' };
 
 /**
  * To be run when bot reloads to capture any changes to servers
